market_backend/requirements.txt
aiofile==3.9.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
attrs==26.1.0
Authlib==1.6.9
bcrypt==5.0.0
beartype==0.22.9
cachetools==7.0.5
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.6
cyclopts==4.10.1
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
ecdsa==0.19.2
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.3
fastapi-mcp==0.4.0
fastmcp==3.2.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
Jinja2==3.1.6
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.27.0
mdurl==0.1.2
more-itertools==11.0.1
openapi-pydantic==0.5.1
opentelemetry-api==1.40.0
packaging==26.0
pathable==0.5.0
platformdirs==4.9.4
py-key-value-aio==0.4.4
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pyperclip==1.11.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sqladmin==0.24.0
SQLAlchemy==2.0.49
sse-starlette==3.3.4
starlette==1.0.0
tomli==2.4.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uncalled-for==0.2.0
urllib3==2.6.3
uvicorn==0.43.0
watchfiles==1.1.1
websockets==16.0
WTForms==3.1.2
zipp==3.23.0
